+++ src/models/data-manager.ts
@@ -67,22 +67,29 @@ export class DataInstanceInternal {
 }
 
 export interface DataManager<Ext, Mod> {
-  capacity: number;
-  count: number;
+  /** PROPERTIES */
 
-  create: () => Ext;
+  /** META */
 
-  get: (id: string) => Ext;
-  getAll: () => Ext[];
-
-  fullReset: (capacity?: number) => void;
-  batchInitializeValue: <K extends keyof Mod>(key: K, array: Mod[K][]) => void;
-  batchInitializeEntites: () => void;
+  // SERIALIZE
   dumpData: () => string;
   loadData: (content: string) => void;
 
+  // EVENTS
   subscribe: (eventName: string, callback: DataManagerEventCallback) => void;
   unsubscribe: (eventName: string, callback: DataManagerEventCallback) => void;
+
+  // FETCH
+  get: (id: string) => Ext;
+  getAll: () => Ext[];
+
+  // MANAGE
+  capacity: number;
+  count: number;
+  create: () => Ext;
+  fullReset: (capacity?: number) => void;
+  batchInitializeValue: <K extends keyof Mod>(key: K, array: Mod[K][]) => void;
+  batchInitializeEntites: () => void;
 }
 
 type DataManagerEventCallback = (...args: any[]) => void;
@@ -112,112 +119,7 @@ export class DataManagerClass<Ext, Int extends Ext & DataInstanceInternal, Mod>
     this.events = {};
   }
 
-  public fullReset(capacity?: number) {
-    this.dataInstances = [];
-    this.dataStore = {} as typeof this.dataStore;
-    this.currentCapacity = capacity ? capacity : this.currentCapacity;
-    this.batchInitializeProperty("publicId");
-  }
-
-  public get capacity(): number {
-    return this.currentCapacity;
-  }
-
-  public get count(): number {
-    return this.dataInstances.length;
-  }
-
-  private resizeStore(alloc?: number): void {
-    throw new Error("Method not implemented yet. Need to allocate more space to create entities.");
-  }
-
-  public create(): Ext {
-    if (this.dataInstances.length >= this.currentCapacity) {
-      this.resizeStore();
-    }
-
-    const newInstId = this.dataInstances.length;
-    const newInstPublicId = crypto.randomUUID();
-    const newInstance = new this.createInstanceCall(newInstId);
-
-    this.dataInstances.push(newInstance);
-    this.dataStore.publicId[newInstId] = newInstPublicId;
-    this.publicToInternal[newInstPublicId] = newInstId;
-
-    return newInstance;
-  }
-
-  public get(id: string): Ext {
-    if (!(id in this.publicToInternal)) {
-      throw new Error("Entity does not exist.");
-    }
-
-    return this.dataInstances[this.publicToInternal[id]];
-  }
-
-  public getPublicId(id: number): string {
-    return this.dataStore["publicId"][id];
-  }
-
-  public getAll(): Ext[] {
-    return [...this.dataInstances];
-  }
-
-  public getSafe(id: string): Ext | undefined {
-    if (!(id in this.publicToInternal)) {
-      return undefined;
-    }
-
-    return this.dataInstances[this.publicToInternal[id]];
-  }
-
-  public getDataValueForInstance(id: number, key: keyof typeof this.dataStore): any {
-    return this.dataStore[key][id];
-  }
-
-  public setDataValueForInstance<K extends keyof typeof this.dataStore>(
-    id: number,
-    key: K,
-    value: (typeof this.dataStore)[K][number]
-  ): boolean {
-    // TODO: do some checks to allow assignment
-    // TODO: throw some exceptions?
-
-    this.dataStore[key][id] = value;
-
-    return true;
-  }
-
-  protected batchInitializeProperty<K extends keyof typeof this.dataStore>(key: K) {
-    this.dataStore = {
-      ...this.dataStore,
-      [key]: new Array<(typeof this.dataStore)[K]>(this.currentCapacity),
-    };
-  }
-
-  public batchInitializeValue<K extends keyof Mod>(key: K, array: Mod[K][]) {
-    if (this.currentCapacity < array.length) {
-      this.resizeStore();
-    }
-
-    const newArray = [
-      ...array,
-      ...new Array<Mod[K]>(Math.min(this.currentCapacity - array.length, 0)),
-    ];
-    this.dataStore = { ...this.dataStore, [key]: newArray };
-  }
-
-  public batchInitializeEntites(): void {
-    if (this.dataInstances.length) {
-      throw new Error("Entities alreaedy exist!");
-    }
-
-    for (let index = 0; index < this.currentCapacity; index++) {
-      this.create();
-    }
-  }
-
-  /** SERIALIZATION */
+  /** SERIALIZE*/
 
   public dumpData(): string {
     return JSON.stringify(this.dataStore);
@@ -262,5 +164,114 @@ export class DataManagerClass<Ext, Int extends Ext & DataInstanceInternal, Mod>
         callback(...args);
       });
     }
+  }
+
+  /** FETCH */
+
+  public get(id: string): Ext {
+    if (!(id in this.publicToInternal)) {
+      throw new Error("Entity does not exist.");
+    }
+
+    return this.dataInstances[this.publicToInternal[id]];
+  }
+
+  public getSafe(id: string): Ext | undefined {
+    if (!(id in this.publicToInternal)) {
+      return undefined;
+    }
+
+    return this.dataInstances[this.publicToInternal[id]];
+  }
+
+  public getPublicId(id: number): string {
+    return this.dataStore["publicId"][id];
+  }
+
+  public getAll(): Ext[] {
+    return [...this.dataInstances];
+  }
+
+  protected getDataValueForInstance(id: number, key: keyof typeof this.dataStore): any {
+    return this.dataStore[key][id];
+  }
+
+  /** MANAGE */
+
+  public get capacity(): number {
+    return this.currentCapacity;
+  }
+
+  public get count(): number {
+    return this.dataInstances.length;
+  }
+
+  public create(): Ext {
+    if (this.dataInstances.length >= this.currentCapacity) {
+      this.resizeStore();
+    }
+
+    const newInstId = this.dataInstances.length;
+    const newInstPublicId = crypto.randomUUID();
+    const newInstance = new this.createInstanceCall(newInstId);
+
+    this.dataInstances.push(newInstance);
+    this.dataStore.publicId[newInstId] = newInstPublicId;
+    this.publicToInternal[newInstPublicId] = newInstId;
+
+    return newInstance;
+  }
+
+  public fullReset(capacity?: number) {
+    this.dataInstances = [];
+    this.dataStore = {} as typeof this.dataStore;
+    this.currentCapacity = capacity ? capacity : this.currentCapacity;
+    this.batchInitializeProperty("publicId");
+  }
+
+  public batchInitializeValue<K extends keyof Mod>(key: K, array: Mod[K][]) {
+    if (this.currentCapacity < array.length) {
+      this.resizeStore();
+    }
+
+    const newArray = [
+      ...array,
+      ...new Array<Mod[K]>(Math.min(this.currentCapacity - array.length, 0)),
+    ];
+    this.dataStore = { ...this.dataStore, [key]: newArray };
+  }
+
+  public batchInitializeEntites(): void {
+    if (this.dataInstances.length) {
+      throw new Error("Entities alreaedy exist!");
+    }
+
+    for (let index = 0; index < this.currentCapacity; index++) {
+      this.create();
+    }
+  }
+
+  private resizeStore(alloc?: number): void {
+    throw new Error("Method not implemented yet. Need to allocate more space to create entities.");
+  }
+
+  protected setDataValueForInstance<K extends keyof typeof this.dataStore>(
+    id: number,
+    key: K,
+    value: (typeof this.dataStore)[K][number]
+  ): boolean {
+    // TODO: do some checks to allow assignment
+    // TODO: throw some exceptions?
+
+    this.dataStore[key][id] = value;
+
+    return true;
+  }
+
+  protected batchInitializeProperty<K extends keyof typeof this.dataStore>(key: K) {
+    this.dataStore = {
+      ...this.dataStore,
+      [key]: new Array<(typeof this.dataStore)[K]>(this.currentCapacity),
+    };
   }
 }
