+++ src/renderer/index.tsx
@@ -20,7 +20,6 @@ if (appElement) {
   render(<App />, appElement);
 }
 
-// Use contextBridge
 window.ipcRenderer.on("main-process-message", (_event, message) => {
   console.log(message);
 });
